Revert "Revert"
This reverts commit e9e67a38474247ea41174d9ca91850ee6fb427b6.

+++ globals.h
@@ -160,7 +160,7 @@ struct nne_solution_paras {
 
 struct gslintegration_paras {
   double nu_edge;
-  double T;
+  float T;
   float *photoion_xs;
 };
